refactor: extract unique violation check into a reusable utility
